Availability Manager for Virtual Machine
This project focuses on the concept of Disaster Management. The
Availability Manager is used to monitor the statuses of the VMs and
hosts. If the VM stops reacting, the program checks if the host is
alive, if the host has failed, it migrates the VM to another host. In
case no host is found, a new host is added to restart the VM.It detects
the live status of the VMs by pinging them, if the ping is successful
then the VM is alive, if it is unsuccessful or if it is timed out, the
VM is considered to be dead. In this case, the vHost is checked for
availability too, if the vHost is found dead, the VMs on that vHost are
migrated to another vHost. If no vHost is available for the migration
to occur, a new vHost is added to the vCenter, and the VMs are migrated
to that (new) vHost